Enabling Cel2 does not need special base fee case (#239)
* Enabling Cel2 does not need special base fee case

We can just continue calculating base fees with the normal EIP-1559
formula. There is not need to special case this block. Without the
special case, we also don't need the test for the special case.

Removing the special case makes it easier to enable Cel2 in other tests
without breaking the tests by unintentionally modifying the base fee.

This change might require a corresponding change in our migration
script for consistency.

* Add comment to consensus/misc/eip1559/eip1559.go

Co-authored-by: piersy <pierspowlesland@gmail.com>

---------

Co-authored-by: piersy <pierspowlesland@gmail.com>