tracing: Show OptimismBaseFeeRecipient in prestate
The OptimismBaseFeeRecipient should show up in prestate tracing results
(both the normal prestate and the diff mode prestate results) if
IsOptimism.
I added one prestate diff test with Optimism turned on to show that it
works correctly. This required adding Random (so that IsMerge is true)
and L1CostFunc to the test block context.