Change header encoding mechanism (#300)
This changes the way that we detect pre-gingerbread headers when encoding.

Closes https://github.com/celo-org/celo-blockchain-planning/issues/588

Previously we would interpret a `gasLimit` of zero to indicate pre-gingerbread headers, since prior to the gingerbread hardfork celo blocks lacked the `gasLimit` field.

This however caused lots of problems with the go tests (existing and ones that continue to be added) that routinely use blocks where gasLimit has not been set.

So now we detect pre-gingerbread blocks when decoding and set a flag on the struct that is used to select the encoding.
New blocks are assumed to be post gingerbread, which is ok because it's not possible to run op-geth without gingerbread set, because we haven't ported across the pre-gingerbread state transition logic.

I was able to revert some test modifications because of this but this didn't fix the tests that broke during the celo11 rebase (skipped in [this commit](https://github.com/celo-org/op-geth/commit/d981efbedcc8c247887aaaeb93dd673d3a48785f)), it seems like those problems were caused by something else.